proof of concept: wasm integration - calls to WebAssembly using embind - separated embind interface from c api - MazeGraph constructor using vector of nodes no longer takes extent as argument - html page which loads wasm and shows that binding of methods works - this commit also contains wip of previous attempts

+++ algolibs/libexhsearch/c_api.h
@@ -18,16 +18,16 @@ extern "C" {
     struct CNode {
         unsigned int node_id;
         unsigned char out_paths; // bitmask. 1, 2, 4, 8 for N, E, S, W, respectively
-        short rotation;
+        int rotation;
     };
 
     struct CGraph {
-        unsigned long long extent;
+        long int extent;
         // The graph is always quadratic.
         // Hence, the given number of nodes is expected to be equal to extent*extent + 1.
         // The nodes specify the row-wise layout of the maze. The last entry is the leftover.
         // Node ids are expected to be unique.
-        unsigned long long num_nodes;
+        unsigned long num_nodes;
         struct CNode * nodes;
     };
